I added here new variable type, because encountered an error in one of my telegram profiles
Necessary execute this SQL-script in your database:

ALTER TABLE user_questions
ALTER COLUMN user_id TYPE BIGINT;

And was added new test

+++ product_assistant/models/schema.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timedelta, UTC
-from sqlalchemy import select, Text, String, Integer, update, DateTime, ForeignKey, func
+from sqlalchemy import select, Text, String, Integer, BigInteger, update, DateTime, ForeignKey, func
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column, Session
 
 
@@ -24,7 +24,7 @@ class UserQuestion(Base):
     __tablename__ = "user_questions"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, nullable=True)
+    user_id: Mapped[int] = mapped_column(BigInteger, nullable=True)
     question_text: Mapped[str] = mapped_column(Text)
     cleaned_text: Mapped[str] = mapped_column(Text, nullable=True)
     product_id: Mapped[int] = mapped_column(Integer, ForeignKey("products.id"), nullable=True)
